pyproject.toml
[project]
name = "test-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosmtplib>=4.0.1",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jwt>=1.3.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
    "websockets>=15.0.1",
]
